Install via Wheel to retain Pipenv scripts support
Why are these changes being introduced:

* In order to have the entrypoint have access to the [scripts] section
  of Pipfile, we seem to need to  use this more complex process for
  installing.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-397

How does this address that need:

* This uses the process we've used for previous Python Docker builds.

Document any side effects to this change:

Multistage Docker builds are more complicated and may be somewhat
confusing to diagnose issues. However, this process seems better than
running a venv inside of Docker which is the other alternative.

A possible third alternative is to not care about the [scripts] section
of Pipfile and change the [entrypoint] to something like
`ENTRYPOINT ["python","submitter/cli.py"]`