Add ADR for metadata JSON file spec
Why these changes are being introduced:
This system needs the metadata files to be consistent so it can parse
them and post the metadata to DSpace when it creates the item. A spec
will allow all submitting systems to create their metadata files using
the same structure.

How this addresses that need:
* Adds ADR 0004 metadata json spec

Side effects of this change:
All submitting systems will need to follow the spec when creating JSON
metadata files.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ETD-427