Adds tests for cli
Why are these changes being introduced:

* Adds tests for previous commits on this branch

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-400

How does this address that need:

* adds moto mocks for sqs and s3 (only sqs is used at this time)
* refactors cli to allow queue names to be passed in to allow for easier
  testing
* updates make tasks to include test and coveralls
* updates ci.yml to have actions run tests and coveralls
* adds initial tests for sample loading and queue reading

Document any side effects to this change:

Default queue names are being set during initialization of the submitter
module. Eventually we'll move to setting those in .env locally and
use an AWS service in production/staging but as that is not yet ready,
this seemed good enough for now.