Add DSpace timeout to configuration
Why these changes are being introduced:
During testing we got several errors due to long server response times
from the DSpace test instance. We know this will continue to be an issue
on the test instance and possibly the production instance as well, so we
need an easy way to set and change the request timeout value without
having to update the code.

How this addresses that need:
* Updates Pipfile to use the latest version of the dspace-python-client
  library, pinning to the main branch instead of a specific release
  since the main branch has the updated timeout param we need.
* Adds a DSPACE_TIMEOUT config variable, set from SSM in stage and prod
  envs and from env variable in dev.
* Uses that DSPACE_TIMEOUT config variable during instantiation of the
  DSpace client.
* Updates the README to clarify use of this setting during development.

Side effects of this change:
SSM params will need to be added for stage and prod environments. This
has already been done for stage (currently set to 120 seconds).

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ETD-473