Handle submission message errors
Why these changes are being introduced:
If a message is submitted to DSS without necessary data, it needs to be
handled appropriately. In cases where the missing data prevents DSS
from writing to a result queue, this should mean crashing the app.
In other cases, it just means writing a result error message to the
result queue without attempting to publish anything to DSpace.

How this addresses that need:
* Adds a config value for expected DSS result queue names
* Adds error classes for missing/invalid result queue (using the above
  config value) and missing message attributes (which are required to
  write a result message)
* Raises the above errors during submission message processing when
  appropriate and lets them bubble up to crash the app with a useful
  error message
* Checks for other submission message errors during message processing
  and creates/writes appropriate result error messages to the specified
  result queue
* Adds tests for the above changes.

Side effects of this change:
New SSM parameter was created for the valid output queues config value.
This will need to be updated whenever a new output queue is added.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ETD-471
* https://mitlibraries.atlassian.net/browse/ETD-444