Add ADR to move specs and create new spec files
Why these changes are being introduced:
Multiple changes to the specifications over time resulted in a confusing
series of ADRs documenting each specification. After discussion, it was
proposed that specs might be better handled as standalone files rather
than in ADRs.

How this addresses that need:
* Adds ADR 008 to move specifications to individual files instead of
  ADRs, superceding previous ADRs 003, 004, 006, and 007
* Creates new spec files for each of the existing specifications using
  the current specification as documented in the ADRs.

Side effects of this change:
The current version of each specification can now be easily found in
the 'docs/specifications' directory of this repo.