Update specifications to match code changes
Why these changes are being introduced:
In the process of testing DSS, some changes were made to the result and
messages in the code that aren't currently represented in the specs. In
addition, clarification was requested around the JSON metadata spec.

How this addresses that need:
* Clarifies language in the metadata JSON spec around general
  requirements, how to handle repeated fields, and adds both "good" and
  "bad" metadata JSON examples
* Clarifies language in the result message spec around general
  requirements and how to handle multiple bitstreams, removes the
  ExceptionMessage field in the error message and adds the
  DSpaceResponse field, and adds both success and error result message
  examples
* Clarifies language in the submission message spec around general
  requirements and how DSS determines primary bitstreams/bitstream
  display order in DSpace, adds submission message example

Side effects of this change:
ETD and Wiley apps may need to be updated to reflect the updated
specifications.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/ETD-458
* https://mitlibraries.atlassian.net/browse/ETD-498