Update Makefile
Update the makefile with the build platform option specified.

The "make dist" command uses a docker build to create the container for the project.
```
dist: ## Build docker container
	docker build -t $(ECR_REGISTRY)/dspacesubmissionservice-stage:latest \
		-t $(ECR_REGISTRY)/dspacesubmissionservice-stage:`git describe --always` \
		-t submitter:latest .
```
However, under certain circumstances, ARM mac users could compile an ARM instance of the container, which doesn't work in AWS.
To fix this, we need to change the build command to:
```
dist: ## Build docker container
	docker build --platform linux/amd64 -t $(ECR_REGISTRY)/dspacesubmissionservice-stage:latest \
		-t $(ECR_REGISTRY)/dspacesubmissionservice-stage:`git describe --always` \
		-t submitter:latest .
```
Which will ensure that the container built is for the right platform.