Automation for AWS Organization
Why these changes are being introduced:
We need to add the deploy/publish automation to this application as we
move it into our AWS Organization.

How this addresses that need:
* Add the three GitHub Actions workflows for dev/stage/prod to
automate publishing the container on PR/Merge/Release
* Remove the stage_ecr_push.yml GitHub Action workflow that is no longer
needed

Side effects of this change:
After this is merged to main, future merges to main will auto-push
containers to Stage-Workloads and future tagged releases will auto-push
containers to Prod-Workloads.