Log unhandled exceptions for debugging
Why these changes are being introduced:

Sentry recently reported, 'Unexpected exception, aborting DSpace
Submission Service processing', which is a prepared logging statement as
part of Submission.submit().  This suggests that a totally unhandled
exception was encountered during submit work.  While somewhat helpful,
this does not give us insight into the underlying exception that caused
it in Sentry or Cloudwatch.

How this addresses that need:
* Changes logger.error to logger.exception to log the underlying
Exception to Sentry with a full traceback
* Though we may still need work to rollback actions taken,
or manage the SQS queue based on an unhandled exception,
this change will provide more information in these situations

Side effects of this change:
* Unhandled exceptions will be fully logged to Sentry

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1035